Moved testUserCanAddProperties from the OrgFolderAddPropChildHealthMetricsTest class to the _OrganizationFolderTest class and removed the OrgFolderAddPropChildHealthMetricsTest class
